--- addon/components/file-uploader/template.hbs
@@ -1,23 +1,0 @@
-{{#if _onUpload}}
-  <div class="upload-progress-contianer">
-    <i class="fa fa-spinner fa-spin"></i> {{file.name}}
-    {{#if useProgress}}
-      <div class="upload-progress-bar">
-        <div class="upload-progress" style="width: {{_progress}}%"></div>
-      </div>
-    {{/if}}
-  </div>
-{{else if hasFile}}
-  {{file.name}}
-  <button class="upf-btn upf-btn--secondary upf-btn--x-small" disabled={{not _isValid}} {{action 'upload'}}>
-    <span>{{t "file_uploader.cta.upload"}}</span>
-  </button>
-  <button class="upf-btn upf-btn--orange--reversed upf-btn--x-small" {{action 'clear'}}>
-    <i class="fa fa-times"></i>
-  </button>
-{{else}}
-  <div class="file-uploader__btn upf-btn upf-btn--secondary upf-btn--x-small">
-    <span>{{text}}</span>
-    {{input-uploader onFile=(action 'onFile')}}
-  </div>
-{{/if}}
