+++ addon/components/http-errors-code.hbs
@@ -1,0 +1,27 @@
+<div class="upf-align--absolute-center margin-left-xx-lg">
+  {{#if (eq this.httpError "404")}}
+    <img src="assets/@upfluence/ember-upf-utils/images/404error.svg" alt={{this.imgDescription}} />
+  {{else}}
+    <img src="assets/@upfluence/ember-upf-utils/images/500error.svg" alt={{this.imgDescription}} />
+  {{/if}}
+
+  <div class="upf-align--center">
+    <p class="text-style-bold text-size-7">{{t (concat "errors." this.httpError ".title")}}</p>
+    <p class="text-color-default-light">
+      {{{t (concat "errors." this.httpError ".subtitle")}}}
+    </p>
+    <div class="error-contact-container">
+      <OSS::Button
+        @skin="primary"
+        @label={{t (concat "errors." this.httpError ".support.button_description")}}
+        {{action "openSupportChannel"}} />
+      <div>
+        {{t (concat "errors." this.httpError ".support.check_description")}}
+        <a href={{this.globalSupportLink}} target="_blank" rel="noopener noreferrer"
+           class="upf-link--reset text-color-bright-purple">
+          {{t (concat "errors." this.httpError ".support.help_description")}}
+        </a>
+      </div>
+    </div>
+  </div>
+</div>
