--- addon/components/http-errors-code/template.hbs
@@ -1,26 +1,0 @@
-<div class="upf-align--absolute-center margin-left-xx-lg">
-  {{#if (eq this.httpError "404")}}
-    <img src="assets/@upfluence/ember-upf-utils/images/404error.svg" alt={{this.imgDescription}} />
-  {{else}}
-    <img src="assets/@upfluence/ember-upf-utils/images/500error.svg" alt={{this.imgDescription}} />
-  {{/if}}
-
-  <div class="upf-align--center">
-    <p class="text-style-bold text-size-7">{{t (concat "errors." this.httpError ".title")}}</p>
-    <p class="text-color-default-light">
-      {{{t (concat "errors." this.httpError ".subtitle")}}}
-    </p>
-    <div class="error-contact-container">
-      <button class="upf-btn upf-btn--primary" {{action "openSupportChannel"}}>
-        {{t (concat "errors." this.httpError ".support.button_description")}}
-      </button>
-      <div>
-        {{t (concat "errors." this.httpError ".support.check_description")}}
-        <a href={{this.globalSupportLink}} target="_blank" rel="noopener noreferrer"
-           class="upf-link--reset text-color-bright-purple">
-          {{t (concat "errors." this.httpError ".support.help_description")}}
-        </a>
-      </div>
-    </div>
-  </div>
-</div>
