+++ addon/components/item-chooser/create-item.hbs
@@ -1,0 +1,6 @@
+<OSS::Button
+  @skin="primary"
+  @size="sm"
+  @label={{{formattedCreateButtonText}}}
+  @loading={{this.isLoading}}
+  {{action "createItem"}} />
