--- addon/components/item-chooser/create-item/template.hbs
@@ -1,3 +1,0 @@
-{{#loading-button slowAction="createItem" class="upf-btn upf-btn--primary upf-btn--x-small"}}
-  {{{formattedCreateButtonText}}}
-{{/loading-button}}
