--- addon/components/ownership-selection-modal/groups/component.js
@@ -1,12 +1,0 @@
-import Component from '@ember/component';
-import layout from './template';
-
-export default Component.extend({
-  layout,
-
-  actions: {
-    updateOwnership(_, defer) {
-      this.saveOwnership(this.entity.ownedBy).finally(defer.resolve);
-    }
-  }
-});
