--- addon/components/ownership-selection-modal/groups/template.hbs
@@ -1,18 +1,0 @@
-<div class="modal-body">
-  <p class="text-size-4 text-color-default-lighter text-center">
-    {{t "ownership_selection.groups.description" entityName=entity.name htmlSafe=true}}
-  </p>
-
-  <label class="margin-top-xx-sm">
-    {{t "ownership_selection.groups.label"}}
-  </label>
-  {{ownership-selection light=true entity=entity}}
-</div>
-
-<div class="modal-footer">
-  {{#loading-button
-    slowAction="updateOwnership" class="upf-btn upf-btn--primary"
-    initiallyDisabled=(not entity.ownedBy)}}
-    {{t "ownership_selection.groups.cta"}}
-  {{/loading-button}}
-</div>
