+++ addon/components/ownership-selection-modal/people.hbs
@@ -1,0 +1,42 @@
+<div class="modal-body">
+  <p class="text-size-4 text-color-default-lighter text-center">
+    {{t "ownership_selection.people.description" entityName=entity.name htmlSafe=true}}
+  </p>
+
+  <label class="margin-top-xx-sm">
+    {{t "ownership_selection.people.label"}}
+  </label>
+  {{#item-chooser
+    multiple=true selection=selectedUsers items=items
+    selectedItemComponent="universal-selection/selected-name"
+    placeholder="Search for a user or email" search=(action "searchEntities")
+    as |opt|}}
+    <div class="team-row team-row--member">
+      <div class="team-row__avatar">
+        {{#if opt.option.avatar_url}}
+          <img src={{opt.option.avatar_url}}
+              class="margin-top-xxx-sm margin-bottom-xxx-sm upf-image upf-image--round-36">
+        {{else}}
+          <div class="upf-image upf-image--round-36 upf-image--alt margin-top-xxx-sm margin-bottom-xxx-sm">
+            <span>
+              {{name-initials (concat opt.option.first_name " " opt.option.last_name)}}
+            </span>
+          </div>
+        {{/if}}
+      </div>
+      <div class="team-row__infos">
+        <div>{{opt.option.first_name}} {{opt.option.last_name}}</div>
+        <div>{{opt.option.email}}</div>
+      </div>
+    </div>
+  {{/item-chooser}}
+</div>
+
+<div class="modal-footer">
+  <OSS::Button
+    @skin="primary"
+    @label={{t "ownership_selection.people.cta"}}
+    @loading={{this.updatingOwnership}}
+    disabled={{noSelectedUsers}}
+    {{action "updateOwnership"}} />
+</div>
