+++ addon/components/publishr/draft-display.hbs
@@ -1,0 +1,7 @@
+<div class="draft-display__content">
+  {{component draftDisplayComponent item=draft.draftAttachable}} 
+</div>
+
+{{#unless contentOnly}}
+  {{publishr/draft-display/comments draft=draft commentable=commentable}}
+{{/unless}}
