+++ addon/components/publishr/draft-display/comments.hbs
@@ -1,0 +1,50 @@
+{{#each draft.draftComments as |draftComment|}}
+  {{#if draftComment.id}}
+    <div class="draft-display__comment {{if draftComment.resolved 'draft-display__comment--resolved'}}">
+      {{#if draftComment.owner.avatar_url}}
+        <img src={{draftComment.owner.avatar_url}} class="upf-image upf-image--round-36">
+      {{else}}
+        <figure class="upf-image upf-image--round-36 upf-image--alt">
+          <span>{{name-initials draftComment.owner.name}}</span>
+        </figure>
+      {{/if}}
+      <div class="margin-left-xxx-sm">
+        <div class="draft-display__comment-header">
+          <div class="owner-infos">
+            <span class="text-style-bold">{{draftComment.owner.name}}</span>
+            <span class="text-color-default-lighter"
+                  {{enable-tooltip title=(format-date draftComment.createdAt 'MMMM Do, YYYY')}}>
+              {{draftComment.relativeCreatedAt}}
+            </span>
+          </div>
+          {{#if (and commentable (eq draftComment.owner.id currentUser.id))}}
+            <a {{action "removeComment" draftComment}} class="remove-comment text-color-error">
+              <i class="fa fa-trash"></i>
+            </a>
+          {{/if}}
+        </div>
+
+        <div class="text-color-default-light">
+          {{draftComment.text}}
+        </div>
+      </div>
+    </div>
+  {{/if}}
+{{/each}}
+
+{{#if commentable}}
+  <div class="margin-top-xxx-sm draft-display__add-comment">
+    {{#input-wrapper}}
+      <label>Comment</label>
+      {{textarea value=commentText class="col-xs-12 upf-input upf-input--textarea"}}
+    {{/input-wrapper}}
+    <div class="upf-align--right">
+      <OSS::Button
+        @skin="primary"
+        @label="Add comment"
+        @loading={{this.savingComment}}
+        disabled={{not this.savableComment}}
+        {{action "addComment"}} />
+    </div>
+  </div>
+{{/if}}
