--- addon/components/publishr/draft-display/template.hbs
@@ -1,7 +1,0 @@
-<div class="draft-display__content">
-  {{component draftDisplayComponent item=draft.draftAttachable}}
-</div>
-
-{{#unless contentOnly}}
-  {{publishr/draft-display/comments draft=draft commentable=commentable}}
-{{/unless}}
