--- addon/components/universal-export/create-item/template.hbs
@@ -1,24 +1,0 @@
-<div class="btn-group dropup">
-  {{#loading-button
-    slowAction="createItem" class="btn upf-btn upf-btn--primary upf-btn--x-small"
-    initiallyDisabled=noCreateableEntities}}
-    {{t "export_influencers.export.create" searchText=publicApi.searchText entity=selectedEntityType}}
-  {{/loading-button}}
-
-  {{#unless noCreateableEntities}}
-    <button type="button" class="btn upf-btn upf-btn--primary upf-btn--x-small dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-      <span class="caret"></span>
-      <span class="sr-only">Toggle Dropdown</span>
-    </button>
-
-    <ul class="dropdown-menu dropdown-menu-right">
-      {{#each createableEntities as |e|}}
-        <li>
-          <a {{action "selectEntityType" e}} class="text-style-capitalize">
-            {{e}}
-          </a>
-        </li>
-      {{/each}}
-    </ul>
-  {{/unless}}
-</div>
