--- app/components/file-uploader/component.js
@@ -1,1 +1,0 @@
-export { default } from '@upfluence/ember-upf-utils/components/file-uploader/component';
