+++ app/components/http-errors-code.js
@@ -1,0 +1,1 @@
+export { default } from '@upfluence/ember-upf-utils/components/http-errors-code';
