--- app/components/http-errors-code/component.js
@@ -1,1 +1,0 @@
-export { default } from '@upfluence/ember-upf-utils/components/http-errors-code/component';
