+++ app/components/item-chooser.js
@@ -1,0 +1,1 @@
+export { default } from '@upfluence/ember-upf-utils/components/item-chooser';
