--- app/components/item-chooser/component.js
@@ -1,1 +1,0 @@
-export { default } from '@upfluence/ember-upf-utils/components/item-chooser/component';
