+++ app/components/item-chooser/create-item.js
@@ -1,0 +1,1 @@
+export { default } from '@upfluence/ember-upf-utils/components/item-chooser/create-item';
