--- app/components/ownership-selection-modal/groups/component.js
@@ -1,1 +1,0 @@
-export { default } from '@upfluence/ember-upf-utils/components/ownership-selection-modal/groups/component';
