+++ app/components/ownership-selection-modal/people.js
@@ -1,0 +1,1 @@
+export { default } from '@upfluence/ember-upf-utils/components/ownership-selection-modal/people';
