+++ app/components/publishr/draft-display/comments.js
@@ -1,0 +1,1 @@
+export { default } from '@upfluence/ember-upf-utils/components/publishr/draft-display/comments';
