--- app/components/publishr/draft-display/comments/component.js
@@ -1,1 +1,0 @@
-export { default } from '@upfluence/ember-upf-utils/components/publishr/draft-display/comments/component';
