--- app/components/publishr/draft-display/component.js
@@ -1,1 +1,0 @@
-export { default } from '@upfluence/ember-upf-utils/components/publishr/draft-display/component';
