+++ app/components/universal-export/create-item.js
@@ -1,0 +1,1 @@
+export { default } from '@upfluence/ember-upf-utils/components/universal-export/create-item';
