--- app/components/universal-export/create-item/component.js
@@ -1,1 +1,0 @@
-export { default } from '@upfluence/ember-upf-utils/components/universal-export/create-item/component';
