+++ app/styles/components/file-uploader.less
@@ -13,6 +13,9 @@
   }
 
   .file-uploader__btn {
+    &:extend(.upf-btn);
+    &:extend(.upf-btn--secondary);
+    &:extend(.upf-btn--sm);
     position: relative;
     overflow: hidden;
   }
